ci: ignore backend
The lock file with the exact version of the module
dependencies should be committed to the source code
until this is done as a workaround the `terraform
init` generates a lock file with the dependencies
of the time of the action run. In this particular
terraform module the workaround cannot work as we
have a backend configured. To overcome this here
we ignore the backend configuration.